requirements.txt
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
pytesseract>=0.3.10
